hababru/requirements.txt
Flask
python-dotenv
requests
PyYAML
markdown
beautifulsoup4
markitdown[all]
pytest
fpdf
python-docx
